change calculation process for anchors

+++ src/cli/utils/anchors/AnchorPeakFinder.java
@@ -9,18 +9,18 @@ import java.util.Set;
 
 public class AnchorPeakFinder {
 
-    public static Set<Anchor> getPeaks(int resolution, Map<Chromosome, double[]> allUpStreamOEProd,
-                                       Map<Chromosome, double[]> allDownStreamOEProd,
-                                       Map<Chromosome, double[]> allBothStreamOEProd) {
+    public static Set<Anchor> getPeaks(int resolution, Map<Chromosome, float[]> allUpStreamOEProd,
+                                       Map<Chromosome, float[]> allDownStreamOEProd,
+                                       Map<Chromosome, float[]> allBothStreamOEProd) {
         Set<Anchor> peaks = new HashSet<>();
         for (Chromosome chrom : allBothStreamOEProd.keySet()) {
-            double[] both = allBothStreamOEProd.get(chrom);
-            double[] up = allUpStreamOEProd.get(chrom);
-            double[] down = allDownStreamOEProd.get(chrom);
+            float[] both = allBothStreamOEProd.get(chrom);
+            float[] up = allUpStreamOEProd.get(chrom);
+            float[] down = allDownStreamOEProd.get(chrom);
 
-            double[] bothSmooth = ConvolutionTools.smooth(both);
-            double[] upSmooth = ConvolutionTools.smooth(up);
-            double[] downSmooth = ConvolutionTools.smooth(down);
+            float[] bothSmooth = ConvolutionTools.smooth(both);
+            float[] upSmooth = ConvolutionTools.smooth(up);
+            float[] downSmooth = ConvolutionTools.smooth(down);
 
             for (int i = 2; i < both.length - 2; i++) {
                 if (elevated5(both, i) && elevated5(bothSmooth, i)) {
@@ -37,7 +37,7 @@ public class AnchorPeakFinder {
         return peaks;
     }
 
-    private static boolean elevated5(double[] data, int i) {
+    private static boolean elevated5(float[] data, int i) {
         return data[i] > 1.1 * data[i - 1]
                 && data[i] > 1.1 * data[i + 1]
                 && data[i - 1] > 1.1 * data[i - 2]
@@ -46,7 +46,7 @@ public class AnchorPeakFinder {
                 && data[i - 2] > 1;
     }
 
-    private static boolean elevated3(double[] data, int i) {
+    private static boolean elevated3(float[] data, int i) {
         return data[i] > 1.1 * data[i - 1]
                 && data[i] > 1.1 * data[i + 1]
                 && data[i + 1] > 1
